+++ R/revdep.R
@@ -64,8 +64,14 @@ print.maintainers <- function(x, ...) {
 #' Run R CMD check on all downstream dependencies.
 #'
 #' Use \code{revdep_check()} to run \code{\link{check_cran}()} on all downstream
-#' dependencies. Summarises the results with \code{revdep_check_summary} and
-#' save logs with \code{revdep_check_save_logs}.
+#' dependencies. Summarises the results with \code{revdep_check_summary()} and
+#' see problems with \code{revdep_check_print_problems()}.
+#'
+#' Revdep checks are resumably - this is very helpful if somethings goes
+#' wrong (like you run out of power or you lose your internet connection) in
+#' the middle of a check. You can resume a partially completed check with
+#' \code{revdep_check_resume()}, or blow away the cached result so you can
+#' start afresh with \code{revdep_check_reset()}.
 #'
 #' @section Check process:
 #' \enumerate{
@@ -88,6 +94,9 @@ print.maintainers <- function(x, ...) {
 #' @inheritParams revdep
 #' @param pkg Path to package. Defaults to current directory.
 #' @inheritParams check_cran
+#' @param check_dir A temporary directory to hold the results of the package
+#'   checks. This should not exist as after the revdep checks complete
+#'   successfully this directory is blown away.
 #' @seealso \code{\link{revdep_maintainers}()} to get a list of all revdep
 #'   maintainers.
 #' @export
@@ -109,36 +118,34 @@ revdep_check <- function(pkg = ".", recursive = FALSE, ignore = NULL,
                          type = getOption("pkgType"),
                          threads = getOption("Ncpus", 1),
                          env_vars = NULL,
-                         check_dir = tempfile("check_cran")) {
+                         check_dir = NULL) {
+
   pkg <- as.package(pkg)
+  if (file.exists(revdep_cache_path(pkg))) {
+    stop("Cache file `revdep/.cache.rds` exists.\n",
+      "Use `revdep_check_resume()` to resume\n",
+      "Use `revdep_check_reset()` to start afresh.",
+      call. = FALSE)
+  }
+
   rule("Reverse dependency checks for ", pkg$package, pad = "=")
 
-  if (!file.exists(libpath))
-    dir.create(libpath)
-  if (!file.exists(srcpath))
-    dir.create(srcpath)
-
-  message(
-    "Installing ", pkg$package, " ", pkg$version,
-    " and dependencies to ", libpath
-  )
-  withr::with_libpaths(libpath, action = "prefix", {
-    install(pkg, reload = FALSE, quiet = TRUE, dependencies = TRUE)
-  })
-  on.exit(remove.packages(pkg$package, libpath), add = TRUE)
+  if (is.null(check_dir)) {
+    check_dir <- file.path(pkg$path, "revdep", "checks")
+    message("Saving check results in `revdep/checks/`")
+  }
+  if (file.exists(check_dir)) {
+    stop("`check_dir()` must not already exist: it is deleted after a successful run",
+      call. = FALSE)
+  }
 
   message("Computing reverse dependencies")
   revdeps <- revdep(pkg$package, recursive = recursive, ignore = ignore,
     bioconductor = bioconductor, dependencies = dependencies)
 
-  env_vars <- c(
-    NOT_CRAN = "false",
-    RGL_USE_NULL = "true",
-    env_vars
-  )
-  show_env_vars(env_vars)
-
-  check_cran(revdeps,
+  # Save arguments and revdeps to a cache
+  cache <- list(
+    pkgs = revdeps,
     libpath = libpath,
     srcpath = srcpath,
     bioconductor = bioconductor,
@@ -147,10 +154,86 @@ revdep_check <- function(pkg = ".", recursive = FALSE, ignore = NULL,
     check_dir = check_dir,
     env_vars = env_vars
   )
+  saveRDS(cache, revdep_cache_path(pkg))
 
-  revdep_check_save(pkg, revdeps, check_dir, libpath)
+  revdep_check_from_cache(pkg, cache)
+}
+
+#' @export
+#' @rdname revdep_check
+revdep_check_resume <- function(pkg = ".") {
+  pkg <- as.package(pkg)
+
+  cache_path <- revdep_cache_path(pkg)
+  if (!file.exists(cache_path)) {
+    message("Previous run completed successfully")
+    return(invisible())
+  }
+
+  cache <- readRDS(cache_path)
+
+  # Don't need to check packages that we've already checked.
+  check_dirs <- dir(cache$check_dir, full.names = TRUE)
+  completed <- file.exists(file.path(check_dirs, "COMPLETE"))
+
+  completed_pkgs <- gsub("\\.Rcheck$", "", basename(check_dirs)[completed])
+  cache$pkgs <- setdiff(cache$pkgs, completed_pkgs)
+
+  revdep_check_from_cache(pkg, cache)
+}
+
+#' @rdname revdep_check
+#' @export
+revdep_check_reset <- function(pkg = ".") {
+  pkg <- as.package(pkg)
+
+  cache_path <- revdep_cache_path(pkg)
+  if (!file.exists(cache_path)) {
+    return(invisible(FALSE))
+  }
+
+  cache <- readRDS(cache_path)
+
+  unlink(cache_path)
+  unlink(cache$check_dir, recursive = TRUE)
+  invisible(TRUE)
+}
+
+revdep_check_from_cache <- function(pkg, cache) {
+  # Install this package into revdep library -----------------------------------
+  if (!file.exists(cache$libpath)) {
+    dir.create(cache$libpath, recursive = TRUE, showWarnings = FALSE)
+  }
+  message(
+    "Installing ", pkg$package, " ", pkg$version,
+    " and dependencies to ", cache$libpath
+  )
+  withr::with_libpaths(cache$libpath, action = "prefix", {
+    install(pkg, reload = FALSE, quiet = TRUE, dependencies = TRUE)
+  })
+  on.exit(remove.packages(pkg$package, cache$libpath), add = TRUE)
+
+  cache$env_vars <- c(
+    NOT_CRAN = "false",
+    RGL_USE_NULL = "true",
+    cache$env_vars
+  )
+  show_env_vars(cache$env_vars)
+
+  do.call(check_cran, cache)
+
+  rule("Saving check results to `revdep/check.rds`")
+  revdep_check_save(pkg, cache$revdeps, cache$check_dir, cache$libpath)
+
+  # Delete cache and check_dir on successful run
+  rule("Cleaning up")
+  revdep_check_reset(pkg)
+  unlink(revdep_cache_path(pkg))
+  unlink(cache$check_dir, recursive = TRUE)
+
   invisible()
 }
+
 
 revdep_check_save <- function(pkg, revdeps, check_path, lib_path) {
   platform <- platform_info()
@@ -165,7 +248,6 @@ revdep_check_save <- function(pkg, revdeps, check_path, lib_path) {
   pkgs <- intersect(pkgs, dir(lib_path))
   dependencies <- package_info(pkgs, libpath = lib_path)
 
-  rule("Saving check results to `revdep/check.rds`")
   out <- list(
     revdeps = revdeps,
     platform = platform,
@@ -193,6 +275,10 @@ parse_package_check <- function(path) {
 
 revdep_check_path <- function(pkg) {
   file.path(pkg$path, "revdep", "checks.rds")
+}
+
+revdep_cache_path <- function(pkg) {
+  file.path(pkg$path, "revdep", ".cache.rds")
 }
 
 check_dirs <- function(path) {